Return URL resolution, redirect support and logo
The `PYROVEN_RETURN_URL` parameter has been removed, instead this url is
inferred from the 'raven_return' django view.

It is now possible to pass a "?next=page" parameter to the 'raven_login'
view which will be used as the target on a successful redirect (this
fits with the standard django auth routines).

The Raven logo has been added.